Prepare eq3btsmart base entity for additional platforms
