+++ homeassistant/components/eq3btsmart/climate.py
@@ -18,19 +18,13 @@ from homeassistant.const import ATTR_TEMPERATURE, PRECISION_HALVES, UnitOfTemper
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.exceptions import ServiceValidationError
 from homeassistant.helpers import device_registry as dr
-from homeassistant.helpers.device_registry import CONNECTION_BLUETOOTH, DeviceInfo
-from homeassistant.helpers.dispatcher import async_dispatcher_connect
+from homeassistant.helpers.device_registry import CONNECTION_BLUETOOTH
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.util import slugify
 
 from . import Eq3ConfigEntry
 from .const import (
-    DEVICE_MODEL,
     EQ_TO_HA_HVAC,
     HA_TO_EQ_HVAC,
-    MANUFACTURER,
-    SIGNAL_THERMOSTAT_CONNECTED,
-    SIGNAL_THERMOSTAT_DISCONNECTED,
     CurrentTemperatureSelector,
     Preset,
     TargetTemperatureSelector,
@@ -75,53 +69,6 @@ class Eq3Climate(Eq3Entity, ClimateEntity):
     _attr_preset_mode: str | None = None
     _target_temperature: float | None = None
 
-    def __init__(self, entry: Eq3ConfigEntry) -> None:
-        """Initialize the climate entity."""
-
-        super().__init__(entry)
-        self._attr_unique_id = dr.format_mac(self._eq3_config.mac_address)
-        self._attr_device_info = DeviceInfo(
-            name=slugify(self._eq3_config.mac_address),
-            manufacturer=MANUFACTURER,
-            model=DEVICE_MODEL,
-            connections={(CONNECTION_BLUETOOTH, self._eq3_config.mac_address)},
-        )
-
-    async def async_added_to_hass(self) -> None:
-        """Run when entity about to be added to hass."""
-
-        self._thermostat.register_update_callback(self._async_on_updated)
-
-        self.async_on_remove(
-            async_dispatcher_connect(
-                self.hass,
-                f"{SIGNAL_THERMOSTAT_DISCONNECTED}_{self._eq3_config.mac_address}",
-                self._async_on_disconnected,
-            )
-        )
-        self.async_on_remove(
-            async_dispatcher_connect(
-                self.hass,
-                f"{SIGNAL_THERMOSTAT_CONNECTED}_{self._eq3_config.mac_address}",
-                self._async_on_connected,
-            )
-        )
-
-    async def async_will_remove_from_hass(self) -> None:
-        """Run when entity will be removed from hass."""
-
-        self._thermostat.unregister_update_callback(self._async_on_updated)
-
-    @callback
-    def _async_on_disconnected(self) -> None:
-        self._attr_available = False
-        self.async_write_ha_state()
-
-    @callback
-    def _async_on_connected(self) -> None:
-        self._attr_available = True
-        self.async_write_ha_state()
-
     @callback
     def _async_on_updated(self) -> None:
         """Handle updated data from the thermostat."""
@@ -132,7 +79,7 @@ class Eq3Climate(Eq3Entity, ClimateEntity):
         if self._thermostat.device_data is not None:
             self._async_on_device_updated()
 
-        self.async_write_ha_state()
+        super()._async_on_updated()
 
     @callback
     def _async_on_status_updated(self) -> None:
